Mailape/requirements.txt
django<2.1
psycopg2<2.8
psycopg2-binary<2.8
django-markdownify==0.3.0
django-crispy-forms==1.7.0

celery<4.2
celery[redis]
django-celery-results<2.0

djangorestframework<3.8

django-extensions==2.1.4